+++ core-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao core bundle change log
 
+### DEV
+
+ * Make services public that we need to access directly.
+
 ### 4.5.0 (2017-12-28)
 
  * Apply the schema filter in the DCA schema provider.